+++ src/DependencyInjection/Configuration.php
@@ -4,24 +4,28 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\FormalizeBundle\DependencyInjection;
 
+use Dbp\Relay\BlobLibrary\Api\BlobApi;
+use Dbp\Relay\BlobLibrary\Api\BlobApiError;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
 class Configuration implements ConfigurationInterface
 {
     public const DATABASE_URL = 'database_url';
-    public const SUBMITTED_FILES_BUCKET_ID = 'submitted_files_bucket_id';
 
+    /**
+     * @throws BlobApiError
+     */
     public function getConfigTreeBuilder(): TreeBuilder
     {
         $treeBuilder = new TreeBuilder('dbp_relay_formalize');
-
-        $treeBuilder->getRootNode()
+        $rootNode = $treeBuilder->getRootNode();
+        $rootNode
             ->children()
-            ->scalarNode(self::DATABASE_URL)->end()
-            ->scalarNode(self::SUBMITTED_FILES_BUCKET_ID)->end()
-            ->end()
-            ->end();
+                ->scalarNode(self::DATABASE_URL)->end()
+             ->end();
+
+        $rootNode->append(BlobApi::getConfigNodeDefinition());
 
         return $treeBuilder;
     }