+++ src/TestUtils/TestUtils.php
@@ -6,6 +6,7 @@ namespace Dbp\Relay\FormalizeBundle\TestUtils;
 
 use Dbp\Relay\AuthorizationBundle\TestUtils\AuthorizationTest;
 use Dbp\Relay\BlobBundle\TestUtils\BlobTestUtils;
+use Dbp\Relay\BlobLibrary\Api\BlobApi;
 use Dbp\Relay\FormalizeBundle\Authorization\AuthorizationService;
 use Dbp\Relay\FormalizeBundle\DependencyInjection\Configuration;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -25,11 +26,9 @@ class TestUtils
 
     public static function getTestConfig(): array
     {
-        return
-            [
-                Configuration::DATABASE_URL => 'sqlite:///:memory:',
-                Configuration::SUBMITTED_FILES_BUCKET_ID => self::FORMALIZE_SUBMITTED_FILES_TEST_BUCKET_ID,
-            ];
+        return array_merge([
+            Configuration::DATABASE_URL => 'sqlite:///:memory:',
+        ], BlobApi::getCustomModeConfig(self::FORMALIZE_SUBMITTED_FILES_TEST_BUCKET_ID));
     }
 
     public static function getBlobTestConfig(): array